+++ src/atoms/index.js
@@ -1,5 +1,6 @@
 // @flow
 
 export { default as Button } from './Button';
+export { default as ButtonQR } from './ButtonQR';
 export { default as Small } from './Small';
 export { default as Text } from './Text';