+++ src/components/BadgerBadge/stories.js
@@ -9,11 +9,14 @@ import BadgerBadge from './BadgerBadge';
 
 import { currencyOptions } from '../../utils/currency-helpers';
 
-const defaultOpReturn = ['0x6d02', 'Hello BadgerBadge']
+const defaultOpReturn = [
+	'0x6d02',
+	'Try out Badger at https://badger.bitcoin.com',
+];
 
 storiesOf('BadgerBadge', module)
 	.add(
-		'default',
+		'all knobs',
 		() => (
 			<BadgerBadge
 				price={number('Price', 0.001)}
@@ -23,13 +26,53 @@ storiesOf('BadgerBadge', module)
 					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
 				)}
 				opReturn={array('OP_RETURN', defaultOpReturn)}
-				tag={text('Button Text', 'Pay')}
+				tag={text('Button Text', 'Badger Pay')}
 				text={text('Top Text', 'Payment Total')}
-				showBrand={boolean('Show Brand', true)}
-				showSatoshis={boolean('Show Satoshis', true)}
+				isRepeatable={boolean('Repeatable payment', false)}
+				repeatTimeout={number('Repeat Timeout (ms)', 4000)}
+				watchAddress={boolean('Watch Address All', true)}
+				showBrand={boolean('Toggle Brand', false)}
+				showSatoshis={boolean('Toggle Satoshis', true)}
+				showQR={boolean('Toggle QR', true)}
+				showBorder={boolean('Toggle Border', false)}
 				successFn={() => console.log('success')}
 				failFn={() => console.log('fail')}
-				
+			/>
+		),
+		{
+			notes:
+				'Badger Badges are perfect for showing the price and Satoshis in a simple clean all in one component.  Default has knobs to experiment with all settings',
+		}
+	)
+	.add(
+		'minimal look',
+		() => (
+			<BadgerBadge
+				amount={0.001}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+				text={''}
+				showSatoshis={false}
+				showQR={boolean('Toggle QR', true)}
+			/>
+		),
+		{
+			notes:
+				'Minimal look of Badge',
+		}
+	)
+	.add(
+		'price in BCH',
+		() => (
+			<BadgerBadge
+				ticker="BCH"
+				amount={number('BCH Amount', 0.0001)}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
 			/>
 		),
 		{
@@ -76,7 +119,26 @@ storiesOf('BadgerBadge', module)
 		}
 	)
 	.add(
-		'optionally satoshis',
+		'toggle QR code',
+		() => (
+			<BadgerBadge
+				price={number('Price', 0.001)}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+				showQR={boolean('Toggle QR', false)}
+				successFn={() => console.log('success')}
+				failFn={() => console.log('fail')}
+			/>
+		),
+		{
+			notes:
+				'Badger Badges are perfect for showing the price and Satoshis in a simple clean all in one component.  Default has knobs to experiment with all settings',
+		}
+	)
+	.add(
+		'toggle satoshis',
 		() => (
 			<BadgerBadge
 				price={0.001}
@@ -94,7 +156,7 @@ storiesOf('BadgerBadge', module)
 		}
 	)
 	.add(
-		'optionally badger info',
+		'toggle badger info',
 		() => (
 			<BadgerBadge
 				price={0.001}
@@ -102,13 +164,47 @@ storiesOf('BadgerBadge', module)
 					'To Address',
 					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
 				)}
-				showBrand={boolean('Badger info', false)}
+				showBrand={boolean('Badger info', true)}
 				successFn={() => console.log('success')}
 				failFn={() => console.log('fail')}
 			/>
 		),
 		{
 			notes: 'Choose to display a link to the Badger homepage',
+		}
+	)
+	.add(
+		'toggle border',
+		() => (
+			<BadgerBadge
+				price={0.001}
+				showBorder={boolean('Toggle Border', true)}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+			/>
+		),
+		{
+			notes: 'Toggle border',
+		}
+	)
+	.add(
+		'repeatable payments',
+		() => (
+			<BadgerBadge
+				price={number('Price', 0.001)}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+				isRepeatable={boolean('Repeatable payment', true)}
+				repeatTimeout={number('Reset Timeout (ms)', 5000)}
+			/>
+		),
+		{
+			notes:
+				'Payments which can happen more than once on a single page visit.  Games for example',
 		}
 	)
 	.add(
@@ -125,7 +221,7 @@ storiesOf('BadgerBadge', module)
 			/>
 		),
 		{
-			notes: 'Custom functions called on Successfull and Failed payments',
+			notes: 'Custom functions called on Successful and Failed payments',
 		}
 	)
 	.add(
@@ -143,6 +239,23 @@ storiesOf('BadgerBadge', module)
 			/>
 		),
 		{
-			notes: 'Change the currency and price',
+			notes: 'Modify the OP_RETURN value when paid with Badger wallet',
+		}
+	)
+	.add(
+		'watch all sources',
+		() => (
+			<BadgerBadge
+				amount={0.0001}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+				watchAddress={boolean('watch Address All', true)}
+			/>
+		),
+		{
+			notes:
+				'if watchAddress is true, the payment will turn to confirmed when the address receives a payment from any source.  Including other people.  This is ideal to use if the payment codes are unique for the checkout.  Not great if the payment address is shared by users.',
 		}
 	);